refactor changes to comply with controller api endpoints
