+++ src/main/java/com/cs544/project/controller/studentView/CourseOfferingController.java
@@ -1,17 +1,14 @@
-package com.cs544.project.controller;
+package com.cs544.project.controller.studentView;
 
 import com.cs544.project.domain.AttendanceRecord;
 import com.cs544.project.domain.CourseOffering;
-import com.cs544.project.domain.CourseRegistration;
 import com.cs544.project.domain.Student;
 import com.cs544.project.dto.StudentCourse;
 import com.cs544.project.exception.CustomNotFoundException;
-import com.cs544.project.repository.CourseRegistrationRepository;
 import com.cs544.project.repository.StudentRepository;
 import com.cs544.project.service.CourseOfferingService;
 import com.cs544.project.service.CourseRegistrationService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.crossstore.ChangeSetPersister;
 import org.springframework.http.HttpStatus;
 import com.cs544.project.domain.Session;
 import com.cs544.project.service.AttendanceRecordService;
@@ -28,8 +25,8 @@ import java.util.Collection;
 import java.util.List;
 
 @RestController
-@RequestMapping("/student-view")
-public class StudentViewController {
+@RequestMapping("/student-view/course-offerings")
+public class CourseOfferingController {
     @Autowired
     private StudentRepository studentRepository;
 
@@ -45,7 +42,7 @@ public class StudentViewController {
     @Autowired
     AttendanceRecordService attendanceRecordService;
 
-    @GetMapping("/course-offerings/{offeringId}")
+    @GetMapping("/{offeringId}")
     public ResponseEntity<?> getCourseOfferingsById(@PathVariable("offeringId") int offeringId)
             throws CustomNotFoundException {
         System.out.println(offeringId);
@@ -53,7 +50,7 @@ public class StudentViewController {
         return ResponseEntity.ok(courseOffering);
     }
 
-    @GetMapping("/course-offerings")
+    @GetMapping()
     public ResponseEntity<?> getCourseOffering() {
         // student should came from the logged in user
         student = studentRepository.findFirstByFirstName("John");
@@ -63,7 +60,7 @@ public class StudentViewController {
         }
         return new ResponseEntity<Collection<StudentCourse>>(studentCourses, HttpStatus.OK);
     }
-    @GetMapping("/course-offerings/{offeringId}/attendance")
+    @GetMapping("/{offeringId}/attendance")
     public ResponseEntity<?> getCourseOfferingAttendance(@PathVariable("offeringId") int offeringId)
             throws CustomNotFoundException {
         // TODO: to be replaced by session-based student (student currently logged in)
