+++ src/main/java/com/cs544/project/controller/sysadmin/StudentController.java
@@ -44,4 +44,5 @@ public class StudentController {
         studentService.deleteStudent(id);
         return ResponseEntity.noContent().build();
     }
+
 }
